fix(ci): python coverage shard line limit — split oversized regression tests
